training/requirements.txt
pandas~=2.2.3
numpy~=2.2.2
joblib~=1.4.2
scikit-learn~=1.6.1
sentence-transformers~=3.4.1
python-dotenv~=1.0.1
groq~=0.18.0